requirements.txt
opencv-python-headless
numpy>=1.19.0
tensorflow==2.12.0
keras>=2.4.0
librosa>=0.8.0
scipy>=1.6.0
pandas
jupyter
notebook
tqdm
opencv-contrib-python-headless
scikit-learn
soundfile
streamlit
psutil
matplotlib
seaborn
pillow
